% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{briggs_pulling_boots}
\alias{briggs_pulling_boots}
\title{Clinical trial costs and outcomes from Briggs et al. article}
\format{A data frame with 2 rows and 7 variables:
\describe{
  \item{group}{placebo or drug}
  \item{cost_mean}{mean cost estimates}
  \item{survival_mean}{mean survival estimates (in years)}
  \item{cost_stdev}{standard deviation of mean cost estimates}
  \item{survival_stdev}{standard deviation of mean survival estimates}
  \item{cost_surv_corr}{correlation between costs and survival}
  \item{n}{number of patients in each group}
}}
\usage{
briggs_pulling_boots
}
\description{
A dataset of the means and standard deviations for a placebo and drug
treatment as used in Briggs, Wonderling, and Mooney. (1997). "Pulling 
cost-effectiveness analysis up by its bootstraps: a non-parametric approach 
to confidence interval estimation." Health Economics, 6, 327-340. Also used
in Briggs and Finn. (1998). "Confidence intervals or surfaces? Uncertainty on
the cost-effectiveness plane." Health Economics, 7, 723-740.
}
\keyword{datasets}
