% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_setup.R
\name{cea_setup.default}
\alias{cea_setup.default}
\title{Create a ceamodel class object}
\usage{
\method{cea_setup}{default}(cst, eff, intv, covt = c(), covt_cst = c(),
  covt_eff = c(), eff_more_better = TRUE, cst_char, eff_char, intv_char,
  covt_char)
}
\arguments{
\item{cst}{A vector of cost values.}

\item{eff}{A vector of effect values.}

\item{intv}{A vector of intervention or program assignment/membership.
cst, eff, and intv must have the same length, greater than one.}

\item{covt}{If provided, a vector (or matrix) of covariate values. covt must
have the same number of rows as cst and eff. The number of columns
is equal to the number of unique covariates.}

\item{covt_cst}{If provided, a vector of integers. The integers correspond
to column numbers of covt for which the covariates will be used for
costs.}

\item{covt_eff}{If provided, a vector of integers. The integers correspond
to column numbers of covt for which the covariates will be used for
effects.}

\item{eff_more_better}{If TRUE, a greater value for effects indicates a
better outcome. If FALSE, a smaller value for effects
indicates a better outcome. Default is TRUE.}

\item{cst_char}{A character string representing the preferred name of the
cost variable (values in cst).}

\item{eff_char}{A character string representing the preferred name of the
effect variable (values in eff).}

\item{intv_char}{A character string representing the preferred name of the
intervention variable (values in intv).}

\item{covt_char}{A character string (or vector of character strings)
representing the preferred name of the covariate
variables. The number of strings provided should equal
the number of columns in covt.}
}
\value{
An object of class ceamodel.
}
\description{
Create a ceamodel class object
}
