% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_setup.R
\name{cea_setup.formula}
\alias{cea_setup.formula}
\title{Create a ceamodel class object}
\usage{
\method{cea_setup}{formula}(formula_cst = formula, formula_eff = formula,
  cea_data = list(), eff_more_better = TRUE)
}
\arguments{
\item{cea_data}{A dataframe containing all of the data for the ceamodel.}

\item{eff_more_better}{If TRUE, a greater value for effects indicates a better
outcome. If FALSE, a smaller value for effects indicates
a better outcome. Default is TRUE.}

\item{cst_char}{A character string representing the name of the cost
variable (values in cst). cst_char must be the name of a column
in the dataframe.}

\item{eff_char}{A character string representing the name of the effect
variable (values in eff). eff_char must be the name of a column
in the dataframe}

\item{intv_char}{A character string representing the preferred name of the
intervention variable (values in intv). intv_char must be the
name of a column in the dataframe}

\item{covt}{If provided, a vector of covariate names. Name(s) in covt must
be columns in the dataframe.}

\item{covt_cst}{If provided, a vector of cost covariate names. Name(s) in
covt_cst must be columns in the dataframe.}

\item{covt_eff}{If provided, a vector of effect covariate values. Name(s) in
covt_eff must be columns in the dataframe.}
}
\value{
An object of class ceamodel.
}
\description{
Create a ceamodel class object
}
