% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_setup.R
\name{cea_setup}
\alias{cea_setup}
\title{Create a ceamodel class object.}
\usage{
cea_setup(x, ...)
}
\arguments{
\item{x}{May be a formula, data frame, or vector of data values.}
}
\value{
An object of class ceamodel.
}
\description{
\code{cea_setup} creates a ceamodel class object that specifies cost and
effect data as well as intervention specifiers for a list of individual 
entities. The model may also specify covariate values to be used in 
estimating incremental costs and effects.
}
