% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_incremental.R
\name{ceamodel_incremental}
\alias{ceamodel_incremental}
\title{Conduct basic incremental analyses using a ceamodel object and setup more
advanced incremental analyses.}
\usage{
ceamodel_incremental(cea_lst = list(), cost_order = TRUE,
  table_print = TRUE)
}
\arguments{
\item{cea_lst}{A ceamodel object created using a ceamodel_setup function.}

\item{cost_order}{If true, the order of options in an ICER table will be by 
increasing average cost. If false, the order of options in an ICER table
will be by increasing average effect.}

\item{table_print}{If true, the ICER table will be printed. If false, the
ICER table will not be printed.}
}
\value{
An object of class ceamodel, with incremental analyses added to the
model
}
\description{
Conduct basic incremental analyses using a ceamodel object and setup more
advanced incremental analyses.
}
\examples{
## Run an incremental analysis
ceamodel <- cea_setup(cost ~ treat, qaly ~ treat, clintrial_cea)
ceamodel <- ceamodel_incremental(ceamodel)
}
