% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceaR.R
\docType{package}
\name{ceaR}
\alias{ceaR}
\alias{ceaR-package}
\title{ceaR: A package for conducting cost-effectivneess analsyes.}
\description{
The ceaR package contains several categories of functions.
}
\section{Setup functions}{

The setup functions ...
}

