% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colorectal_screening}
\alias{colorectal_screening}
\title{Colorectal cancer screening cost-effectiveness analysis}
\format{A dataframe with 11 observations and 3 variables:
\describe{
  \item{strategy}{Screening strategy}
  \item{costs}{Average costs in 2008 Canadian dollars}
  \item{qalys}{Average quality-adjusted life-years following screening}
}}
\usage{
colorectal_screening
}
\description{
Results from a cost-utility analysis using a Markov model of screening for 
colorectal cancer. Strategies compared inlcude guaiac-based
fecal occult blood test (FOBT) or fecal immunochemical test (FIT) annually, 
fecal DNA every 3 years, flexible sigmoidoscopy or computed
tomographic colonography every 5 years, and colonoscopy every 10 years.
Published in Heitman, Hilsden et al. (2010). "Colorectal Cancer Screening for 
Average-Risk North Americans: An Economic Evaluation." PLoS Med 7(11): 
e1000370. doi:10.1371/journal.pmed.1000370
}
\keyword{datasets}
