% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_other.R
\name{create_formula}
\alias{create_formula}
\title{Creates a formula to be used in ceaR functions}
\usage{
create_formula(dep, intv, covt)
}
\arguments{
\item{dep}{Dependent variable}

\item{intv}{Intervention variable}

\item{covt}{Covariate variables}
}
\description{
Creates a formula for used in various ceaR functions based on a dependent
variable, intervention variable, and covariates. Each argument should be 
passed as a character vector and the object returned is of class formula.
}
