% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_ceplane_plots.R
\name{create_icer_plot}
\alias{create_icer_plot}
\title{Print an ICER plot}
\usage{
create_icer_plot(slopes_vec, intercept_vec, xlim_vec = c(), ylim_vec = c(),
  icer_marker = c(), title_lab = "", x_lab = "Incremental Effects",
  y_lab = "Incremental Costs", save.plot = FALSE,
  file_name = "create_icer_plot.jpg", shade_df = NA)
}
\description{
Print an ICER plot
}
