% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icer_table.R
\name{create_icer_table}
\alias{create_icer_table}
\title{Create an ICER table based on a set of average cost and effect values by
a set of unique intervention variables.}
\usage{
create_icer_table(cst_vec, eff_vec, intv_names, eff_more_better = TRUE,
  cost_order = TRUE, table_print = TRUE)
}
\arguments{
\item{cst_vec}{A vector of average cost values the same length as intv_names.}

\item{eff_vec}{A vector of average effect values the same length as 
intv_names.}

\item{intv_names}{A vector of character strings representing intervention
names for each unique intervention to include a row in the ICER table.}

\item{eff_more_better}{If TRUE, a greater value for effects indicates a
better outcome. If FALSE, a smaller value for effects
indicates a better outcome. Default is TRUE.}

\item{cost_order}{If true, the order of options in an ICER table will be by 
increasing average cost. If false, the order of options in an ICER table
will be by increasing average effect.}

\item{table_print}{If true, the ICER table will be printed to the Console.
If false, priting of the ICER table to the Console will be suppressed.}
}
\value{
Here
}
\description{
Create an ICER table based on a set of average cost and effect values by
a set of unique intervention variables.
}
