% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_ceplane_plots.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Print a scatterplot}
\usage{
create_scatter_plot(cea_lst = ceamodel(), title_lab = "",
  x_lab = "Incremental Effects", y_lab = "Incremental Costs",
  file_name = "create_scatter_plot.jpg", incremental = TRUE)
}
\description{
Print a scatterplot
}
