% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1}
\alias{example1}
\title{Cost-effectiveness example dataset}
\format{A dataframe with 3 observations and 3 variables:
\describe{
  \item{Intervention}{indicators of intervention}
  \item{costs}{mean costs of intervention}
  \item{lys_gained}{mean life-years gained of each intervention}
}}
\usage{
example1
}
\description{
Small dataset of three generic interventions with mean costs and life-years 
gained.
}
\keyword{datasets}
