% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2}
\alias{example2}
\title{Cost-effectiveness example dataset}
\format{A dataframe with 5 observations and 3 variables:
\describe{
  \item{Intervention}{indicators of intervention}
  \item{costs}{mean costs of intervention}
  \item{qalys}{mean qalys of each intervention}
}}
\usage{
example2
}
\description{
Small dataset of three generic interventions with mean costs and qalys.
}
\keyword{datasets}
