% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_ceplane_plots.R
\name{fieller_ci_plot.ceamodel}
\alias{fieller_ci_plot.ceamodel}
\title{Print a Fieller confidence interval plot, class ceamodel method}
\usage{
\method{fieller_ci_plot}{ceamodel}(cea_lst,
  title_lab = "Bow Tie ICER Confidence Region\\n                                               (Fieller's theorem)",
  x_lab = "Incremental Effects", y_lab = "Incremental Costs",
  save.plot = FALSE, file_name = "create_fieller_ci_plot.jpg")
}
\description{
Print a Fieller confidence interval plot, class ceamodel method
}
