% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_other.R
\name{get_zval}
\alias{get_zval}
\title{Returns a Z-value based on confidence level}
\usage{
get_zval(ci_level = 95)
}
\arguments{
\item{ci_level}{Level of confidence. The default value is 95, representing
a 95% level of confidence or an alpha of 0.05.}
}
\value{
Z-value for the standard normal distribution
}
\description{
Returns a Z-value for use in confidence interval construction based on
a given level of confidence. The level of confidence is passed as an integer
representing the level of confidence.
}
