% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icer_analysis.R
\name{icer_analysis}
\alias{icer_analysis}
\title{Statistical analysis on an ICER}
\usage{
icer_analysis(cea_lst = list(), intv1, intv2, fieller_plot = FALSE,
  pvalues = FALSE, inb = FALSE, boot = FALSE, ci_level = 95)
}
\arguments{
\item{cea_lst}{A ceamodel class object created using a ceamodel_setup function.
If a ceaicer object is not included (from ceamodel_incremental), this will
be run first before an analysis of the ICER.}

\item{intv1}{Reference to one of the conditions specified in the intervention
variable of cea_lst}

\item{fieller_plot}{If TRUE, create and save a Fieller confidence interval
plot for the ICER}

\item{inb}{If TRUE, run an incremental net-benefit analysis}

\item{boot}{If TRUE, run a bootstrap analysis}

\item{ci_level, }{Percent level for the confidence interval, default at 95%
confidence interval.}

\item{intv1}{Reference to one of the conditions specified in the intervention
variable of cea_lst, should be different from intv1}
}
\description{
Runs a statistical analysis on an ICER value of interest following an
incremental analysis. The ICER of interest is selected by specifying the
two conditions of interest with intv1 and intv2.
}
