% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inb.R
\name{inb}
\alias{inb}
\title{Run an incremental net benefit analysis}
\usage{
inb(ceamodel = list(), intv1 = NA, intv2 = NA, n_lam = 101,
  lam_upper = NA)
}
\arguments{
\item{ceamodel}{Object of class ceamodel}

\item{intv1}{Reference to one of the conditions specified in the intervention
variable of cea_lst}

\item{n_lam}{Number of lambda values to run in analysis}

\item{lam_upper}{Upper limit of lambda values to run in analysis}

\item{intv1}{Reference to one of the conditions specified in the intervention
variable of cea_lst, should be different from intv1}
}
\value{
An object of class XXX
}
\description{
must have an object from ceamodel_incremental run first
if icer_analysis has been run then intv1 and intv2 exist and will be pulled
}
