% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_class_funcs.R
\name{plot.ceamodel}
\alias{plot.ceamodel}
\title{Default plotting model for a ceamodel object}
\usage{
\method{plot}{ceamodel}(x, save.plot = FALSE, ...)
}
\arguments{
\item{x}{A ceamodel object}
}
\description{
Default plotting model for a ceamodel object
}
