% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_class_funcs.R
\name{print.ceamodel}
\alias{print.ceamodel}
\title{Print information regarding a ceamodel object}
\usage{
\method{print}{ceamodel}(x, ...)
}
\arguments{
\item{x}{A ceamodel object}
}
\description{
Print information regarding a ceamodel object
}
