% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_class_funcs.R
\name{print.icermodel}
\alias{print.icermodel}
\title{Print information regarding an icermodel object}
\usage{
\method{print}{icermodel}(x, ...)
}
\arguments{
\item{x}{An icermodel object}
}
\description{
Print information regarding an icermodel object
}
