% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_class_funcs.R
\name{print.icertable}
\alias{print.icertable}
\title{Print an icertable object}
\usage{
\method{print}{icertable}(x, ...)
}
\arguments{
\item{x}{An icertable object}
}
\description{
Print an icertable object
}
