% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prostate_simulation}
\alias{prostate_simulation}
\title{Prostate cancer screening simulations study based on ERSPC data}
\format{A data frame with 6 rows and 14 variables:
\describe{
  \item{strategy}{generic indicator of screening strategy}
  \item{description}{description of screening strategy}
  \item{num_tests}{number of screening tests per 1000 men}
  \item{num_men}{number of unique men screened}
  \item{cancers_diag}{number of cancers diagnosed per 1000 men}
  \item{cancers_screened}{number of cancers detected through screening per
  1000 men}
  \item{cancers_overdiagnosed}{number of cancers diagnosed through screening 
  that would not have caused problems without screening per 1000 men}
  \item{death_prostate_cancer}{number of deaths from prostate cancer per
  1000 men}
  \item{ly_gained}{life-years gained (relative to no screening) per 1000 men}
  \item{qaly_gained}{qalys gained (relative to no screening) per 1000 men}
}}
\usage{
prostate_simulation
}
\description{
Results from Heijnsdijk, Carvalho, Auvinen et al. (2015). "Cost-effectiveness
of prostate cancer screening: a simulation study based on ERSPC data." 
JNCI, 107(1), dju366. The results of the European Randomized Study of 
Screening for Prostate Cancer (ERSPC) trial showed a statistically 
significant prostate cancer mortality reduction for the men screened in the 
intervention arm. Based on this data, the authors developed a 
microsimulation model to test a number of screening scenarios. All numbers
are given per 1000 simulated men.
}
\keyword{datasets}
