% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quitline_cea}
\alias{quitline_cea}
\title{Cost-effectiveness of a state tobacco quitline}
\format{A dataframe with 6 observations and 8 variables:
\describe{
  \item{intervention}{intervention strategy including counselling intensity
  and inclusion or not of NRT}
  \item{abstinence_6mos}{proportion of the sample abstinent at 6 months 
  follow-up}
  \item{abstinence_12mos}{proportion of the sample abstinent at 12 months 
  follow-up}
  \item{cost_mean}{mean cost per participant}
  \item{cost_sd}{standard deviation of the mean cost per participant}
  \item{n}{number of participants}
  \item{intensity}{indicator of intervention intensity}
  \item{nrt}{indicator of NRT inclusion in intervention}
}}
\usage{
quitline_cea
}
\description{
Examines the cost effectiveness of offering callers single session versus 
multisession counselling, with or without free nicotine patches (nrt).
Published in Hollis, McAfee, Fellows et al. (2007). "The effectiveness and
cost-effectiveness of telephone counselling and the nicotine patch in a state
tobacco quitline." Tobacco Control, 16(Suppl 1), i53-i59. Brief, moderate, 
and intensive refer to the length and frequency of counselling sessions.
NRT referrs to nicotine replacement therapy in the form of patches. The
outcome refers to the proportion of the sample that was abstinent at two
different time points.
}
\keyword{datasets}
