% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sureg.R
\name{sureg}
\alias{sureg}
\title{Conduct a seemingly unrelated regression analysis for incremental cost-
effectiveness analyses.}
\usage{
sureg(mf_lst = list(), intv1 = NA, intv2 = NA)
}
\arguments{
\item{mf_lst}{A cea model object with incremental analysis specified.}

\item{intv1}{Here}

\item{intv2}{Here}
}
\value{
An sureg model object
}
\description{
Conduct a seemingly unrelated regression analysis for incremental cost-
effectiveness analyses.
}
