% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twopart.R
\name{twopart_model}
\alias{twopart_model}
\title{Conduct a two-part model for excess zeros in cost models}
\usage{
twopart_model(formula, family = gaussian, data)
}
\description{
Conduct a two-part model for excess zeros in cost models
}
