% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceamodel_setup.R
\name{cea_setup}
\alias{cea_setup}
\alias{cea_setup.default}
\alias{cea_setup.data.frame}
\alias{cea_setup.formula}
\title{Create a ceamodel class object.}
\usage{
cea_setup(x, ...)

\method{cea_setup}{default}(cst, eff, intv, covt = c(), covt_cst = c(),
  covt_eff = c(), eff_more_better = TRUE, cst_char, eff_char, intv_char,
  covt_char, call.txt, incremental = TRUE, cost_order = TRUE,
  cst_type = "gaussian", eff_type = "gaussian")

\method{cea_setup}{data.frame}(cea_data = list(), cst_char, eff_char,
  intv_char, covt_char_vec = c(), covt_cst = c(), covt_eff = c(),
  eff_more_better = TRUE, incremental = TRUE, cost_order = TRUE,
  cst_type = "gaussian", eff_type = "gaussian")

\method{cea_setup}{formula}(formula_cea = formula, intv, cea_data = list(),
  eff_more_better = TRUE, incremental = TRUE, cost_order = TRUE,
  cst_type = "gaussian", eff_type = "gaussian")
}
\arguments{
\item{x}{an R object.}

\item{cst}{A vector of cost values.}

\item{eff}{A vector of effect values.}

\item{intv}{A vector of intervention or program assignment/membership.
The vectors cst, eff, and intv must have the same length, greater 
than one.}

\item{covt}{If provided, a vector (or matrix) of covariate values. covt must
have the same number of rows as cst and eff. The number of columns
is equal to the number of unique covariates.}

\item{covt_cst}{If provided, a vector of integers. The integers correspond
to column numbers of covt for which the covariates will be used for
costs.}

\item{covt_eff}{If provided, a vector of integers. The integers correspond
to column numbers of covt for which the covariates will be used for
effects.}

\item{eff_more_better}{If TRUE, a greater value for effects indicates a
better outcome. If FALSE, a smaller value for effects
indicates a better outcome. Default is TRUE.}

\item{cst_char}{A character string representing the preferred name of the
cost variable (values in cst).}

\item{eff_char}{A character string representing the preferred name of the
effect variable (values in eff).}

\item{intv_char}{A character string representing the preferred name of the
intervention variable (values in intv).}

\item{covt_char}{A character string (or vector of character strings)
representing the preferred name of the covariate
variables. The number of strings provided should equal
the number of columns in covt.}

\item{call.txt}{Only supplied when one of the cea_setup methods calls this
default function.}

\item{incremental}{Defaults to TRUE and runs incremental analysis upon setup.}

\item{cost_order}{If true, the order of options in an ICER table will be by 
increasing average cost. If false, the order of options in an ICER table
will be by increasing average effect.}

\item{cst_type}{Defaults to linear regression (gaussian) but can be set
to any family object available to glm.}

\item{eff_type}{Defaults to linear regression (gaussian) but can be set
to any family object available to glm.}
}
\value{
An object of class "ceamodel".
}
\description{
Creates a ceamodel class object that specifies cost and
effect data as well as intervention specifiers for a list of individual 
entities. The model may also specify covariate values to be used in 
estimating incremental costs and effects.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default S3 method

\item \code{data.frame}: S3 method for class 'data.frame'

\item \code{formula}: S3 method for class 'formula'
}}

\examples{
## Data Frame examples: basic, male as covariate for effects only
ceamodel <- cea_setup(clintrial_cea, "cost", "qaly", "treat")
ceamodel <- cea_setup(clintrial_cea, "cost", "qaly", "treat", 
covt_eff = "male")

## Formula examples: basic, male as covariate for effects only
ceamodel <- cea_setup(cost | qaly ~ 1, "treat", clintrial_cea)
ceamodel <- cea_setup(cost | qaly ~ 1 | male, "treat", clintrial_cea)
}
